requirements.txt
autopep8==1.5.5
cycler==0.10.0
graphviz==0.16
kiwisolver==1.3.1
matplotlib==3.3.4
neat-python==0.92
numpy==1.19.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
